+++ projects/fhi-angular-components/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 1.0.0
 
 > Jun 1, 2023